pvr: Implement NULL jobs with drmSyncobjTransfer() in the DRM backend
Ideally, we should stop doing this consolidation at the generic level
and instead pass vk_sync_{wait,signal} arrays to the winsys job submit
methods. To help with out fence consolidation in the DRM driver, we can
add a CONSOLIDATE flag to drm_pvr_sync_op::flags, which would allow us
to merge output fences coming from different jobs into a single syncobj
point. If input syncobj consolidation is needed for the pvrsrvkm backend
(we don't need that for the DRM driver, since we can pass an array of
input syncobjs to wait on) it should be done there, not in the generic
submission path.

Anyway, the winsys job submission abstraction has been designed around
this job consolidation approach, and refactoring things to match the
proposed model is not trivial. So let's go for a simpler approach here,
and use the SYNCOBJ_TRANSFER ioctl and a temporary timeline syncobj to
do the consolidation without relying on a driver specific ioctl() that
would otherwise not be required.

Signed-off-by: Boris Brezillon <boris.brezillon@collabora.com>