pvr: Get the driver to allocate vtxin and coeff regs
This commits adds code to allocate the vertex input regs and
coefficient registers and passes that info back to the compiler.

This commit also sets up the compute pipeline's state based
on that info and mock state that would be provided by the
compiler.

For now the old hard coding path is kept as to not break currently
operational demos/tests.

Signed-off-by: Karmjit Mahil <Karmjit.Mahil@imgtec.com>